+++ wgpu-hal/examples/ray-traced-triangle/main.rs
@@ -429,7 +429,7 @@ impl<A: hal::Api> Example<A> {
         let indices_size_in_bytes = indices.len() * 4;
 
         let vertices_buffer = unsafe {
-            let vertices_buffer = device
+            let (vertices_buffer, _) = device
                 .create_buffer(&hal::BufferDescriptor {
                     label: Some("vertices buffer"),
                     size: vertices_size_in_bytes as u64,
@@ -455,7 +455,7 @@ impl<A: hal::Api> Example<A> {
 
         let indices_buffer = if index_buffer {
             unsafe {
-                let indices_buffer = device
+                let (indices_buffer, _) = device
                     .create_buffer(&hal::BufferDescriptor {
                         label: Some("indices buffer"),
                         size: indices_size_in_bytes as u64,
@@ -565,7 +565,7 @@ impl<A: hal::Api> Example<A> {
         let instances_buffer_size = instances.len() * size_of::<AccelerationStructureInstance>();
 
         let instances_buffer = unsafe {
-            let instances_buffer = device
+            let (instances_buffer, _) = device
                 .create_buffer(&hal::BufferDescriptor {
                     label: Some("instances_buffer"),
                     size: instances_buffer_size as u64,
@@ -631,7 +631,7 @@ impl<A: hal::Api> Example<A> {
         let uniforms_size = size_of::<Uniforms>();
 
         let uniform_buffer = unsafe {
-            let uniform_buffer = device
+            let (uniform_buffer, _) = device
                 .create_buffer(&hal::BufferDescriptor {
                     label: Some("uniform buffer"),
                     size: uniforms_size as u64,
@@ -722,7 +722,7 @@ impl<A: hal::Api> Example<A> {
             unsafe { device.create_bind_group(&group_desc).unwrap() }
         };
 
-        let scratch_buffer = unsafe {
+        let (scratch_buffer, _) = unsafe {
             device
                 .create_buffer(&hal::BufferDescriptor {
                     label: Some("scratch buffer"),